Full API

Full public API

Flow operator API

The basic building block of this package is the Flows.Flow object, a discrete approximation of the flow of a dynamical system. Here is a list of possible constructors.

Flows.flow — Function.

flow(g, m::AbstractMethod, ts::AbstractTimeStepping)

Construct an object of type Flow, representing the numerical dicretisation of the time-forward flow operator associated to the vector field g, using the integration method m, with time stepping provided by ts. This method should be used with an explicit integration method.

flow(g, A, m::AbstractMethod, ts::AbstractTimeStepping)

Construct a flow operator associated to the vector field defined by a linear component A and a nonlinear part g. An implicit-explicit integration method m should be provided.

flow(g::Coupled{N}, m::AbstractMethod, ts::AbstractTimeStepping) where {N}

Construct a flow operator associated to the composite vector field g, using a default call dependency structure, i.e. where the elements of g satisfy the calling interface:

g[1](t, u[1], dudt[1])
g[2](t, u[1], dudt[1], u[2], dudt[2]) ...
g[N](t, u[1], dudt[1], u[N], dudt[N])

See Flows.jl Call Dependencies for more details on how to specify custom call dependencies. This method should be used with an explicit integrator.

flow(g::Coupled{N}, spec::CallDependency{N}, m::AbstractMethod, ts::AbstractTimeStepping) where {N}

Similar to the method without spec, but specifying a custom call dependency structure.

flow(g::Coupled{N}, A::Coupled{N}, m::AbstractMethod, ts::AbstractTimeStepping) where {N}

Similar to previous methods, but also provide the linear part of the dynamical system. This method should be used with an implicit-explicit integrator.

flow(g::Coupled{N}, A::Coupled{N}, spec::CallDependency{N}, m::AbstractMethod, ts::AbstractTimeStepping) where {N}

Similar to previous methods, but provide a custom call dependency structure. This method should be used with an implicit-explicit integrator.

Objects of type Flow satisfy a callable interface, with additional arguments possible.

Flows.Flow — Type.

(I::Flow)(x, span::NTuple{2, Real})

Map x at time span[1] to the later time span[2].

The object x is modified in place. The argument x shoule be of a type compatible to that used to create the integration method object for the Flow object I, since the integration method contains preallocated elements used to perform the integration step.

(I::Flow)(x, span::NTuple{2, Real}, m::AbstractMonitor)

Map x at time span[1] to the later time span[2], filling the monitor obejct m along the way. See Flow.jl Monitor objects for more details on how to define and use Monitor objects.

(I::Flow)(x, span::NTuple{2, Real}, c::AbstractStageCache)

Map x at time span[1] to the later time span[2], filling the stage cache obejct c along the way. See Flow.jl Stage Caches for more details on how to define and use AbstractStageCache objects.

(I::Flow)(x, span::NTuple{2, Real}, s::AbstractStorage)

Map x at time span[1] to the later time span[2], filling the storage object s along the way. This method is used primarily to fill a storage object with the results of a nonlinear simulation, where the storage s can be subsequently used for the linearised systems. See Flow.jl Storages for more details on how to define and use AbstractStorage objects.

(I::Flow{TimeStepFromCache})(x, c::AbstractStageCache, m::Union{Nothing, <:AbstractMonitor}=nothing)

Map x forward/backward over a time span defined by the stage cache object c, filling the monitor object m along the way. This method is primarily used to integrate linearised equations forward/backward, so that the nonlinear and linearised methods are discretely consistent. See Flow.jl Stage Caches for more details on how to define and use AbstractStorage objects.

(I::Flow{TimeStepFromStorage})(x, s::AbstractStorage, span::NTuple{2, Real}, m::Union{Nothing, <:AbstractMonitor}=nothing)

Map x forward/backward over a time span (span[1], span[2]) using the nonlinear trajectory stored in s to drive linearised equations. An additional Monitor object m can be filled along the way. The monitor object is fed with elements that are similar to x. This method can be used to integrate forward or adjoint equations in a way that is not discretely consistent.

Monitor API

Flows.Monitor — Type.

Monitor(x, f::Base.Callable=identity, store::S=RAMStorage(f(x)); oneevery::Int=1, savebetween::Tuple{Real, Real}=(-Inf, Inf), sizehint::Int=0)

Construct a Monitor object to record one observable quantity along a trajectory.

The argument x is an object of the same type used to represent the system's state, while f is a callable object or function that calculates the observable from the state. In other words, the quantity f(x) is monitored along a trajectory, and stored in store, which defaults to a RAMStorage object. One sample every onevery samples is stored.

If required, only samples at times falling in the range specified by savebetween are stored. Specifying the number of samples stored with the sizehint keyword argument may increase performance.

A Monitor object can then be passed as an additional argument to a Flows.Flow object.

See also reset!, times and samples.

Flows.reset! — Function.

reset!(mon::Monitor, sizehint::Int=0)

Reset the internal storage of a Monitor object mon.

Flows.times — Function.

times(mon::Monitor)

Return the times at which samples of the observable have been stored. This is most typically after each time step, in addition to the initial condition. The type of the returned object depend on the internal storage. For RAMStorage storages, this is a standard Vector.

Flows.samples — Function.

samples(mon::Monitor)

Return samples of the observable that have been stored during a trajectory. This is most typically after each time step, in addition to the initial condition. The type of the returned object depend on the internal storage. For RAMStorage storages, this is a standard Vector.

Coupled API

Flows.Coupled — Type.

Coupled{N, ARGS<:NTuple{N, Any}} <: AbstractVector{Any}

Couple together N objects and store them internally in a Julia ntuple of size N.

Coupledobjects are immutable, but their elements must be mutable, allowing the content to be modified. The individual elements of aCoupledobject can be accessed using the standard indexing notation, but the elements cannot be changed, i.e. there is noBase.setindex!defined forCoupled` objects.

Flows.couple — Function.

couple(args...)

Create a Coupled object from the sequence of arguments args.

Flows.couplecopy — Function.

couplecopy(N::Int, x)

Couple N copies of x together, created using Base.deepcopy.

Base.getindex — Function.

Base.getindex(x::Coupled, i::Int)

Return the i-th element of a Coupled object.

Base.similar — Function.

Base.similar(x::Coupled)

Call Base.similar on the elements of x and couple them together in a new object.

Base.size — Function.

Base.size(::Coupled{N}) where {N}

Return (N, ), the size of a Coupled{N} object, coupling N elements together.

Storage API

Flows.RAMStorage — Type.

empty

Flows.period — Function.

period(rs::RAMStorage)

Return the period of the data, or 0 if the data in non periodic.

Flows.isperiodic — Function.

isperiodic(rs::RAMStorage)

Return true if the data stored in rs represent a periodic signal.

Flows.timespan — Function.

timespan(rs::RAMStorage)

Return a 2-tuple with the first and last times stored.

Integration methods API

Flows.RK4 — Type.

RK4(x::X, mode::AbstractMode=NormalMode())

Constructs a RK4 integration scheme object for integration with mode mode.

Flows.CNRK2 — Type.

CNRK2(x::X, mode::AbstractMode=NormalMode())

Constructs a CNRK2 integration scheme object for integration with mode mode.

Flows.CB3R2R2 — Type.

CB3R2R2(x::X, mode::AbstractMode=NormalMode())

Constructs a CB3R2R2 integration scheme object for integration with mode mode.

Flows.CB3R2R3c — Type.

CB3R2R3c(x::X, mode::AbstractMode=NormalMode())

Constructs a CB3R2R3c integration scheme object for integration with mode mode.

Flows.CB3R2R3e — Type.

CB3R2R3e(x::X, mode::AbstractMode=NormalMode())

Constructs a CB3R2R3e integration scheme object for integration with mode mode.

Flows.CB4R3R4 — Type.

CB4R3R4(x::X, mode::AbstractMode=NormalMode())

Constructs a CB4R3R4 integration scheme object for integration with mode mode.

Flows.ImcA! — Function.

ImcA!(A, c::Real, y, z)

Return z that solves the linear problem (I - c*A)*z = y, where c is a scalar, A a linear operator and I is the identity.

To use the IMEX schemes in this package, user should add methods to this function for their custom types, using an efficient implementation. A default implementation when A is of type LinearAlgebra.Diagonal and y and z is provided by this package.

Notes

The name of this function should be read "I-minus-see-A".

Time Stepping API

Flows.TimeStepConstant — Type.

TimeStepConstant(Δt::Real)

Specify that integration should be performed with constant time step Δt.

Flows.TimeStepFromStorage — Type.

TimeStepFromStorage(Δt::Real)

Specify that integration should be performed with constant time step Δt, and that an AbstractStorage will be required.

Flows.AbstractTimeStepFromHook — Type.

AbstractTimeStepFromHook

Abstract type for time stepping schemes where the time step is determined at runtime.

See Flows.jl Adaptive time stepping